Add monster rejection chance handling. Update combats.txt to specify it. Slight format change to how to specify even and odd ascension monsters in combats.txt. Wild seahorse stats from wiki, fix for Bat quest tracking. Experimental, encounter chances in Bat Hole are varied based on Bat quest and number of turns spent in Screambat areas.
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@14695 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/persistence/AdventureQueueDatabase.java
@@ -327,7 +327,7 @@ public class AdventureQueueDatabase
 		String zone = data.getZone();
 		RollingLinkedList zoneQueue = COMBAT_QUEUE.get( zone );
 
-		int denominator = data.totalWeighting();
+		double denominator = data.totalWeighting();
 
 		// without queue effects the result is just numerator/denominator.
 		if ( zoneQueue == null )
